Refactor to provide array of numbers, not objects
This was important because I use cart data to fetch from the API, but
when I make an array of ids from the array of cart objects, this array
is new. This caused an infinite loop:
component uses hook -> hook refreshes component -> component use hook ..
So now an array of IDs itself is returned so that React knows that
the dependencies haven't changed.

I also split the pushing of changes to subscribers in CartData so that
when only the count of items changes, the array of IDs is not updated.